requirements.txt
Flask>=2.2.0
Werkzeug>=2.0.0
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
mysql-connector-python==8.0.30